+++ go/tasks/plugins/array/awsbatch/transformer.go
@@ -2,6 +2,7 @@ package awsbatch
 
 import (
 	"context"
+	"github.com/lyft/flyteplugins/go/tasks/pluginmachinery/coreutils"
 	"sort"
 	"time"
 
@@ -19,8 +20,6 @@ import (
 	"github.com/aws/aws-sdk-go/service/batch"
 	"github.com/lyft/flyteidl/gen/pb-go/flyteidl/core"
 	idlCore "github.com/lyft/flyteidl/gen/pb-go/flyteidl/core"
-	"github.com/lyft/flyteplugins/go/tasks/pluginmachinery/utils"
-
 	"github.com/lyft/flyteplugins/go/tasks/errors"
 	pluginCore "github.com/lyft/flyteplugins/go/tasks/pluginmachinery/core"
 	v1 "k8s.io/api/core/v1"
@@ -66,12 +65,12 @@ func FlyteTaskToBatchInput(ctx context.Context, tCtx pluginCore.TaskExecutionCon
 		return nil, errors.Errorf(errors.BadTaskSpecification, "config[%v] is missing", DynamicTaskQueueKey)
 	}
 
-	cmd, err := utils.ReplaceTemplateCommandArgs(ctx, taskTemplate.GetContainer().GetCommand(), tCtx.InputReader(), tCtx.OutputWriter())
+	cmd, err := coreutils.ReplaceTemplateCommandArgs(ctx, tCtx.TaskExecutionMetadata(), taskTemplate.GetContainer().GetCommand(), tCtx.InputReader(), tCtx.OutputWriter())
 	if err != nil {
 		return nil, err
 	}
 
-	args, err := utils.ReplaceTemplateCommandArgs(ctx, taskTemplate.GetContainer().GetArgs(),
+	args, err := coreutils.ReplaceTemplateCommandArgs(ctx, tCtx.TaskExecutionMetadata(), taskTemplate.GetContainer().GetArgs(),
 		arrayJobInputReader{tCtx.InputReader()}, tCtx.OutputWriter())
 	taskTemplate.GetContainer().GetEnv()
 	if err != nil {
