+++ go/tasks/plugins/presto/execution_state.go
@@ -2,6 +2,7 @@ package presto
 
 import (
 	"context"
+	"github.com/lyft/flyteplugins/go/tasks/pluginmachinery/coreutils"
 
 	"github.com/lyft/flyteplugins/go/tasks/pluginmachinery/ioutils"
 
@@ -219,7 +220,7 @@ func GetQueryInfo(ctx context.Context, tCtx core.TaskExecutionContext) (string, 
 		return "", "", "", "", err
 	}
 
-	outputs, err := utils.ReplaceTemplateCommandArgs(ctx, []string{
+	outputs, err := coreutils.ReplaceTemplateCommandArgs(ctx, tCtx.TaskExecutionMetadata(), []string{
 		prestoQuery.RoutingGroup,
 		prestoQuery.Catalog,
 		prestoQuery.Schema,
